feat concurrent/thread: move native threads tests from thread subdirectory to native_thread
